refactor : ViewModel 구조 변경
